Allow plugins to instert customised html and js in forms
Adds to moderate_post.html and new.html the possiblity for a plugin
to add easily some html to it.
To add specific fields or behaviour for your module in new or
moderate_post you should create a folder in templates/plugins
with the name of your plugin. Inside that folder you should create
a file named moderate.html (or/and new.html if you want to add
some stuff when creating a new post). Same if you want to add
some personlised scripts. Make a file in the same directory
(templates/plugins/plugin_name/) with the name moderate.js
(or/and new.js).

If you want to add this feature to new templates you can juste add the
same modifications than in new.html and moderate_post.html. Add
then in the plugin folder of your module the file with the
corresponding name. Don't forget to make sure that you pass the channel
in argument when calling the template.

The place where the plugin stuff will be added is for now below
data_until, it can maybe be moved above.

Co-authored-by: Maxmaw <maxime.mawait@student.uclouvain.be>
Co-authored-by: Sami <sami.bosch@student.uclouvain.be>
Co-authored-by: bensim602 <benjamin.simon@student.uclouvain.be>
Co-authored-by: sluyters <arthur.sluyters@student.uclouvain.be>
Co-authored-by: Gui <guillaume.gheysen@student.uclouvain.be>
Co-authored-by: TGLuis <luis.tascon@student.uclouvain.be>
Co-authored-by: tanguyl <tanguy.losseau@student.uclouvain.be>